+++ src/index.js
@@ -1,7 +1,13 @@
 'use strict'
 /* global self */
 
+const isIPFS = require('is-ipfs')
+const CID = require('cids')
 const multiaddr = require('multiaddr')
+const multibase = require('multibase')
+const multihash = require('multihashes')
+const PeerId = require('peer-id')
+const PeerInfo = require('peer-info')
 const loadCommands = require('./utils/load-commands')
 const getConfig = require('./utils/default-config')
 const sendRequest = require('./utils/send-request')
@@ -39,7 +45,6 @@ function ipfsClient (hostOrMultiaddr, port, opts) {
   const requestAPI = sendRequest(config)
   const cmds = loadCommands(requestAPI, config)
   cmds.send = requestAPI
-  cmds.Buffer = Buffer // Added buffer in types (this should be removed once a breaking change is release)
 
   return cmds
 }
@@ -54,3 +59,5 @@ function toHostAndPort (multiaddr) {
 }
 
 module.exports = ipfsClient
+
+Object.assign(module.exports, { isIPFS, Buffer, CID, multiaddr, multibase, multihash, PeerId, PeerInfo })
