+++ test/interface.spec.js
@@ -282,8 +282,4 @@ describe('interface-ipfs-core tests', () => {
       }
     }
   }))
-
-  tests.types(defaultCommonFactory)
-
-  tests.util(defaultCommonFactory, { skip: { reason: 'FIXME currently failing' } })
 })
